+++ DBSimple_EF/DBSimple_EF.csproj
@@ -55,7 +55,9 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Data\Artist.cs" />
     <Compile Include="Data\SongsDB.cs" />
+    <Compile Include="Data\Track.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
