+++ DBSimple_EF/Data/Artist.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
+
+namespace DBSimple_EF.Data
+{
+    [Table("Artists")]
+    public class Artist
+    {
+        public int Id { get; set; }
+
+        [Required]
+        public string Name { get; set; }
+
+        public DateTime Birthday { get; set; }
+    }
+}
